+++ src/hud/dialoguebox.cpp
@@ -66,7 +66,7 @@ void DialogueBox::Render(DialoguePageMode mode, bool hasName, float nameWidth,
   if (mode == DPM_ADV && hasName) {
     if (HasSpeakerPortraits) {
       // Draw Face
-      for (int i = 0; i < SpeakerPortraits.size(); i++) {
+      for (size_t i = 0; i < SpeakerPortraits.size(); i++) {
         int bufId = ScrWork[SW_FACE1SURF + i];
         SpeakerPortraits[bufId].UpdateState(i);
